+++ draftsman/_factorio_version.py
@@ -1,4 +1,4 @@
 # _factorio_version.py
 
-__factorio_version__ = "1.1.78.0"
-__factorio_version_info__ = (1, 1, 78, 0)
+__factorio_version__ = "1.1.81.0"
+__factorio_version_info__ = (1, 1, 81, 0)